♻️ refactor: Muda o tipo de `basePrice` de `double` para `BigDecimal` na classe `Garage` e no record `GarageConfigDTO`

+++ src/main/java/com/Estapar/EstaparParkingSystem/parkingSystem/application/api/dto/GarageConfigDTO.java
@@ -2,9 +2,11 @@ package com.Estapar.EstaparParkingSystem.parkingSystem.application.api.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
+
 public record GarageConfigDTO(
         String sector,
-        double basePrice,
+        BigDecimal basePrice,
 
         @JsonProperty("max_capacity")
         int maxCapacity
